requirements.txt
appdirs==1.4.4
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
black==19.10b0
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
Flask==1.1.2
Flask-MySQL==1.5.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mypy-extensions==0.4.3
numpy==1.19.2
pandas==1.1.3
pathspec==0.7.0
pip==21.0.1
psycopg2==2.8.6
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.3.17
scikit-learn==0.24.1
scipy==1.5.2
setuptools==52.0.0.post20210125
six==1.15.0
soupsieve==2.2.1
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wheel==0.36.2
wincertstore==0.2
